backend/requirements.txt
acres==0.3.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
Brotli==1.1.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
chardet==5.2.0
charset-normalizer==3.4.0
ci-info==0.3.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
configobj==5.0.9
configparser==7.2.0
constantly==23.10.4
contourpy==1.3.1
cron-descriptor==1.4.5
cryptography==43.0.3
cssselect2==0.7.0
cycler==0.12.1
daphne==4.1.2
defusedxml==0.8.0rc2
Deprecated==1.2.18
distlib==0.3.9
dj-rest-auth==7.0.1
Django==5.2
django-allauth==65.7.0
django-celery-beat==2.7.0
django-cors-headers==4.6.0
django-filter==24.3
django-prometheus==2.3.1
django-redis==5.4.0
django-rest-marshmallow==4.0.2
django-timezone-field==7.0
Django-Verify-Email==3.0.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djongo==1.3.7
djoser==2.3.1
dnspython==2.7.0
drf-yasg==1.21.8
efficientnet_pytorch==0.7.1
email_validator==2.2.0
etelemetry==0.3.1
fastapi==0.115.8
filelock==3.16.1
fonttools==4.55.3
fsspec==2025.2.0
googleapis-common-protos==1.59.1
grpcio==1.71.0
h11==0.14.0
httplib2==0.22.0
httptools==0.6.4
huggingface-hub==0.30.2
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.6.1
incremental==24.7.2
inflection==0.5.1
iniconfig==2.0.0
isodate==0.6.1
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.7
kombu==5.4.2
looseversion==1.3.0
lxml==5.3.2
MarkupSafe==3.0.2
marshmallow==4.0.0
matplotlib==3.9.3
mpmath==1.3.0
munch==4.0.0
mysqlclient==2.2.7
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
pretrainedmodels==0.7.4
prometheus_client==0.21.1
prompt_toolkit==3.0.48
protobuf==4.25.7
prov==2.0.1
psycopg2-binary==2.9.10
puremagic==1.28
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydot==3.0.4
pydyf==0.11.0
PyJWT==2.9.0
pymongo==3.11.4
PyMuPDF==1.25.5
pyOpenSSL==25.0.0
pyotp==2.9.0
pyparsing==3.2.0
pyphen==0.17.2
pytest==8.3.3
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.20
python3-openid==3.2.0
pytz==2024.2
pyxnat==1.6.3
PyYAML==6.0.2
qrcode==8.1
rdflib==6.3.2
redis==5.2.0
regex==2024.11.6
reportlab==4.3.1
requests==2.32.3
requests-oauthlib==2.0.0
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
segmentation_models_pytorch==0.4.0
sentence-transformers==4.1.0
sentry-sdk==2.22.0
service-identity==24.2.0
setuptools==75.8.0
simplejson==3.20.1
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
SQLAlchemy==2.0.37
sqlparse==0.5.3
starlette==0.45.3
sympy==1.13.1
threadpoolctl==3.6.0
thrift==0.21.0
timm==1.0.14
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.21.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
traits==7.0.2
transformers==4.51.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.34.0
uvloop==0.21.0
vine==5.1.0
virtualenv==20.27.1
openai>=1.0.0
